Updated portrait gallery sorting
It was pointed out that something has changed on the Jekyll builds of
GitHub and the sorting of the project gallery on the main page is no
longer controlled by filename. I updated the index.html file to
re-enable this functionality. To illustrate this change, I renamed
Project 3 to show that it shows up as the first file. Additionally I
added a permalink tag to the Project 3 MD file to show how you could
still provide nice URLs while using sorting of the *.md files to control
the ordering.

See more here:
http://stackoverflow.com/questions/9053066/sorted-navigation-menu-with-jekyll-and-liquid